+++ examples/data-sources/appleads_campaign_report/data-source.tf
@@ -1,0 +1,4 @@
+data "appleads_campaign_report" "us" {
+  start_time = "2026-09-01"
+  end_time   = "2026-09-14"
+}